Move Rotation class to its own header
