+++ LibCarla/source/carla/geom/Rotation.h
@@ -1,0 +1,47 @@
+// Copyright (c) 2017 Computer Vision Center (CVC) at the Universitat Autonoma
+// de Barcelona (UAB).
+//
+// This work is licensed under the terms of the MIT license.
+// For a copy, see <https://opensource.org/licenses/MIT>.
+
+#pragma once
+
+#include "carla/MsgPack.h"
+
+#ifdef LIBCARLA_INCLUDED_FROM_UE4
+#  include "Math/Rotator.h"
+#endif // LIBCARLA_INCLUDED_FROM_UE4
+
+namespace carla {
+namespace geom {
+
+  class Rotation {
+  public:
+
+    Rotation() = default;
+
+    Rotation(float p, float y, float r)
+      : pitch(p),
+        yaw(y),
+        roll(r) {}
+
+    float pitch = 0.0f;
+    float yaw = 0.0f;
+    float roll = 0.0f;
+
+#ifdef LIBCARLA_INCLUDED_FROM_UE4
+
+    Rotation(const FRotator &rotator)
+      : Rotation(rotator.Pitch, rotator.Yaw, rotator.Roll) {}
+
+    operator FRotator() const {
+      return FRotator{pitch, yaw, roll};
+    }
+
+#endif // LIBCARLA_INCLUDED_FROM_UE4
+
+    MSGPACK_DEFINE_ARRAY(pitch, yaw, roll);
+  };
+
+} // namespace geom
+} // namespace carla
